move models, fix typo
