--- package.json
@@ -1,15 +1,0 @@
-{
-  "name": "ironhack-cohort-tools-project",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "mongoose": "^8.2.3"
-  }
-}
