+++ server/package.json
@@ -18,9 +18,11 @@
   },
   "homepage": "https://github.com/ironhack-labs/cohort-tools-project#readme",
   "dependencies": {
+    "bson": "^6.5.0",
     "cookie-parser": "^1.4.6",
     "cors": "^2.8.5",
     "express": "^4.18.2",
+    "mongoose": "^8.2.3",
     "morgan": "^1.10.0",
     "nodemon": "^3.0.1"
   }
